+++ src/client/app/models/paragraph-content.ts
@@ -1,5 +1,7 @@
+import {ParagraphContentType} from "./paragraph-content-type.enum";
 export class ParagraphContent {
 
-  type: string = "ParagraphContent";
+  type: ParagraphContentType = ParagraphContentType.Label;
+  //type: string  = "ParagraphContent";
 
 }
