+++ src/server/models/formation.ts
@@ -1,115 +1,14 @@
 import * as _ from "lodash";
 import db from "./db";
+import {ParagraphType} from "./eParagraphType";
+import {ParagraphContentType} from "./eParagraphContentType";
+import {IFormation, modelIFormation} from "./iFormation";
 import Mongoose = require("mongoose");
+import {IParagraph} from "./iParagraph";
 var Schema = Mongoose.Schema;
 var debug = require('debug')('server:model:formation');
 
-class IFormation {
-
-  // The name of the formation
-  name: string;
-
-  // The description
-  description: string;
-
-  // The average note (choosen by the users)
-  note: number;
-  // The note count (to add a new note to the average)
-  noteCount: number;
-
-  // Is it favorite of the user (won't be in the model because coming from the user)
-  isFavorite: boolean;
-
-  // Is this interresting for the user between 0 and 1 (won't be in the model because coming from the user)
-  interest: number;
-
-  // When the user start and end the course (won't be in the model because coming from the user)
-  dateFollowed: Date;
-  dateFollowedEnd: Date;
-
-  // Percentage he already followed (won't be in the model because coming from the user)
-  percentFollowed: number;
-
-  created: Date;
-  updated: Date;
-
-
-  /**
-   * Constructor
-   * @param mongoose.Document<IFormation>
-   */
-  constructor(document: {}) {
-    _.merge(this, document);
-  }
-}
-
-interface IFormationModel extends IFormation, Mongoose.Document {
-}
-
-
-/**
- * MongooseSchema
- * @type {"mongoose".Schema}
- * @private
- */
-var _schema: Mongoose.Schema = new Mongoose.Schema({
-    name: {
-      type: String,
-      require: true
-    },
-    description: {
-      type: String,
-      require: false
-    },
-    note: {
-      type: Number,
-      require: false
-    },
-    noteCount: {
-      type: Number,
-      default: 0
-    },
-    created: {
-      type: Date,
-      default: Date.now
-    },
-    updated: {
-      type: Date,
-      default: Date.now
-    }
-
-  })
-    .pre('save', function (next) {
-      //console.log("pre save");
-      this.updated = new Date();
-      next();
-    })
-  // .post('init', function (doc) {
-  //   console.log('%s has been initialized from the db', doc['_id']);
-  // })
-  // .post('count', function (doc) {
-  //   console.log('%s has been initialized from the db', doc['_id']);
-  // })
-  // .post('validate', function (doc) {
-  //   console.log('%s has been validated (but not saved yet)', doc['_id']);
-  // })
-  // .post('save', function (doc) {
-  //   console.log('%s has been saved', doc['_id']);
-  // })
-  // .post('remove', function (doc) {
-  //   console.log('%s has been removed', doc['_id']);
-  // })
-  ;
-
-
-/**
- * Mongoose.Model
- * @type {Model<IParagraphModel>}
- * @private
- */
-var _model = Mongoose.model < IFormationModel >('Formation', _schema);
-
-class Formation extends IFormation {
+export default class Formation extends IFormation {
 
   /**
    * Constructor
@@ -117,7 +16,7 @@ class Formation extends IFormation {
    */
   constructor(document: {}) {
     super(document);
-    _model.on('error', function (err) {
+    modelIFormation.on('error', function (err) {
       debug("Error : " + err);
     });
   }
@@ -126,7 +25,7 @@ class Formation extends IFormation {
   static count(): Promise<number> {
     debug("count");
     return new Promise<number>((resolve, reject) => {
-      _model.count(
+      modelIFormation.count(
         {},
         (err, count) => {
           //debug("count " + err + " " + count);
@@ -151,7 +50,7 @@ class Formation extends IFormation {
     return new Promise < IFormation[] >((resolve, reject) => {
 
       // Do the search
-      _model.find({})
+      modelIFormation.find({})
         .exec()
         .then(
           formations => {
@@ -171,10 +70,11 @@ class Formation extends IFormation {
 
   static findById(id: string): Promise < Formation > {
     return new Promise < IFormation >((resolve, reject) => {
-      _model.findById(id)
+      modelIFormation.findById(id)
         .exec()
         .then(
           formation => {
+            debug(formation);
             resolve(formation);
           },
           err => {
@@ -189,7 +89,7 @@ class Formation extends IFormation {
       // debug("updateOrCreate id:" + formation["_id"]);
       if (formation["_id"]) {
         formation.updated = new Date();
-        _model.findByIdAndUpdate(formation["_id"], formation)
+        modelIFormation.findByIdAndUpdate(formation["_id"], formation)
           .exec()
           .then(
             paragraph => {
@@ -199,7 +99,7 @@ class Formation extends IFormation {
               reject(err);
             });
       } else {
-        _model.create(formation)
+        modelIFormation.create(formation)
           .then(
             paragraph => {
               resolve(paragraph);
@@ -215,13 +115,11 @@ class Formation extends IFormation {
 
 }
 
-export = Formation
-
 
 // Init database if it's empty
 Formation.count()
   .then(count => {
-    //debug("init then");
+    debug("init then");
     if (count === 0) {
       var formations: {}[] = [
         {
@@ -231,6 +129,276 @@ Formation.count()
           noteCount: 4,
           created: new Date('2016-08-12T00:00:00'),
           updated: new Date('2016-08-12T00:10:00'),
+          parts: [
+          {
+            title: "Week 1 : Big Data, how it works ?",
+            parts: [
+              {
+                title: "What is Big Data ? (from business point of view)",
+                parts: null,
+                contents: [
+
+                  new IParagraph({
+                      type: ParagraphType.MarkDown,
+                      content: [
+                        '*quelques mots* ou  _quelques mots_\n\n**plus important** ou __également important__',
+                        'Mon texte `code` fin de mon texte',
+                        '    Première ligne de code\n    Deuxième ligne',
+                        '> Ce texte apparaîtra dans un élément HTML blockquote.',
+                        '* Pommes\n* Poires\n    * Sous élément avec au moins quatre espaces devant.',
+                        '1. mon premier\n2. mon deuxième',
+                        '# un titre de premier niveau\n#### un titre de quatrième niveau\n\n',
+                        'Titre de niveau 1\n=====================\n\nTitre de niveau 2\n--------------------'
+                      ]
+                    }
+                  ),
+                  new IParagraph({
+                      type: ParagraphType.MarkDown,
+                      content: [
+                        '+	this is a list item\nindented with tabs\n\n' +
+                        '+   this is a list item\n' +
+                        'indented with spaces\n\n' +
+                        'Code:\n\n	this code block is indented by one tab\n\n' +
+                        'And: \n\n		this code block is indented by two tabs\n\n' +
+                        'And: \n\n+	this is an example list item\n		indented with tabs\n\n' +
+                        '+   this is an example list item\n	    indented with spaces'
+                      ]
+                    }
+                  ),
+                  new IParagraph({
+                      type: ParagraphType.MarkDown,
+                      content: [
+                        '\`\`\`javascript\n  var s = "JavaScript syntax highlighting";\n  alert(s);\n\`\`\`',
+                        '\`\`\`python\n  s = "Python syntax highlighting"\n  print s\n\`\`\`',
+                        '\`\`\`\n  No language indicated, so no syntax highlighting.\n  But let\'s throw in a <b>tag</b>.\n\`\`\`'
+                      ]
+                    }
+                  ),
+                  new IParagraph({
+                      type: ParagraphType.Form,
+                      content: [
+                        {
+                          type: ParagraphContentType.Label,
+                          label: 'This is the title of the question'
+                        },
+                        {
+                          type: ParagraphContentType.Radio,
+                          name: 'F239712893F323AB35',
+                          label: 'Is response **True** ?\n\nSecond line',
+                          value: '1'
+                        },
+                        {
+                          type: ParagraphContentType.Radio,
+                          name: 'F239712893F323AB35',
+                          label: 'Is response **False** ?',
+                          value: '2'
+                        },
+                        {
+                          type: ParagraphContentType.Radio,
+                          name: 'F239712893F323AB35',
+                          label: 'Or **Neither**',
+                          value: '3'
+                        },
+                      ],
+                      userChoice: undefined,
+                      userCheckOK: undefined,
+                      userCheckCount: 0,
+                      maxCheckCount: 2
+                    }
+                  ),
+                  new IParagraph({
+                      type: ParagraphType.Form,
+                      content: [
+                        {
+                          type: ParagraphContentType.Label,
+                          label: 'This is the title of the question'
+                        },
+                        {
+                          type: ParagraphContentType.Checkbox,
+                          name: 'FC56E98',
+                          label: 'Is response **True** ?\n\nSecond line',
+                          value: '1'
+                        },
+                        {
+                          type: ParagraphContentType.Checkbox,
+                          name: 'FC56E98',
+                          label: 'Is response **False** ?',
+                          value: '2'
+                        },
+                        {
+                          type: ParagraphContentType.Checkbox,
+                          name: 'FC56E98',
+                          label: 'Or **Neither**',
+                          value: '3'
+                        },
+                      ],
+                      userChoice: undefined,
+                      userCheckOK: undefined,
+                      userCheckCount: 0,
+                      maxCheckCount: 3
+                    }
+                  ),
+                  new IParagraph({
+                      type: ParagraphType.Form,
+                      content: [
+                        {
+                          type: ParagraphContentType.Label,
+                          label: 'This is the title of the question'
+                        },
+                        {
+                          type: ParagraphContentType.Text,
+                          name: 'fec4637',
+                          label: 'What do you think ?\n\nSecond line',
+                          size: 20
+                        },
+                      ],
+                      userChoice: undefined,
+                      userCheckOK: undefined,
+                      userCheckCount: 0,
+                      maxCheckCount: 3
+                    }
+                  ),
+                  new IParagraph({
+                      type: ParagraphType.Form,
+                      content: [
+                        {
+                          type: ParagraphContentType.Label,
+                          label: 'This is the title of the question'
+                        },
+                        {
+                          type: ParagraphContentType.Radio,
+                          name: 'F239712893F323AB35',
+                          label: 'Is response **True** ?\n\nSecond line',
+                          value: '1'
+                        },
+                        {
+                          type: ParagraphContentType.Radio,
+                          name: 'F239712893F323AB35',
+                          label: 'Is response **False** ?',
+                          value: '2'
+                        },
+                        {
+                          type: ParagraphContentType.Radio,
+                          name: 'F239712893F323AB35',
+                          label: 'Or **Neither**',
+                          value: '3'
+                        },
+                      ],
+                      userChoice: 2,
+                      userCheckOK: false,
+                      userCheckCount: 2,
+                      maxCheckCount: 2,
+                      answer: 1
+                    }
+                  ),
+                  new IParagraph({
+                      type: ParagraphType.Form,
+                      content: [
+                        {
+                          type: ParagraphContentType.Label,
+                          label: 'This is the title of the question'
+                        },
+                        {
+                          type: ParagraphContentType.Checkbox,
+                          name: 'FC56E98',
+                          label: 'Is response **True** ?\n\nSecond line',
+                          value: '1'
+                        },
+                        {
+                          type: ParagraphContentType.Checkbox,
+                          name: 'FC56E98',
+                          label: 'Is response **False** ?',
+                          value: '2'
+                        },
+                        {
+                          type: ParagraphContentType.Checkbox,
+                          name: 'FC56E98',
+                          label: 'Or **Neither**',
+                          value: '3'
+                        },
+                      ],
+                      userChoice: ['1', '2'],
+                      userCheckOK: false,
+                      userCheckCount: 3,
+                      maxCheckCount: 3,
+                      answer: ['1', '3']
+                    }
+                  ),
+                  new IParagraph({
+                      type: ParagraphType.Form,
+                      content: [
+                        {
+                          type: ParagraphContentType.Label,
+                          label: 'This is the title of the question'
+                        },
+                        {
+                          type: ParagraphContentType.Text,
+                          name: 'fec4637',
+                          label: 'What do you think ?\n\nSecond line',
+                          size: 20
+                        },
+                      ],
+                      userChoice: 'rgeg',
+                      userCheckOK: false,
+                      userCheckCount: 3,
+                      maxCheckCount: 3,
+                      answer: 'sdkfsd sdg'
+                    }
+                  ),
+                  new IParagraph({
+                      type: ParagraphType.Form,
+                      content: [
+                        {
+                          type: ParagraphContentType.Label,
+                          label: 'This is the title of the question'
+                        },
+                        {
+                          type: ParagraphContentType.Text,
+                          name: 'fec4637',
+                          label: 'What do you think ?',
+                          size: 20
+                        },
+                      ],
+                      userChoice: 'rgeg',
+                      userCheckOK: false,
+                      userCheckCount: 3,
+                      maxCheckCount: 3,
+                      answer: 'sdkfsd sdg'
+                    }
+                  ),
+                ]
+              },
+              {
+                title: "What is Big Data ? (from developer point of view)",
+                parts: [],
+                contents: null
+              },
+
+            ],
+            contents: null
+          },
+          {
+            title: "Week 2 : Let's try something",
+            parts: [
+              {
+                title: "Prerequisite : A little bit of unix",
+                parts: null,
+                contents: null
+              },
+              {
+                title: "Let's code",
+                parts: null,
+                contents: null
+              },
+            ],
+            contents: null
+          },
+          {
+            title: "Week 3 : A little bit deeper",
+            parts: null,
+            contents: null
+          },
+        ]
         },
         {
           name: "What's new in JDK8",
