+++ src/server/models/iParagraphContent.ts
@@ -1,0 +1,5 @@
+export class IParagraphContent {
+
+  type: string = "IParagraphContent";
+
+}
