+++ src/server/models/iUserChoices.ts
@@ -1,0 +1,15 @@
+
+class UserChoices {
+  userCheckCount: number;
+  userCheckOK: boolean;
+  userChoice: any;
+  created: Date;
+  updated: Date;
+
+  constructor() {
+
+  }
+}
+
+
+export = UserChoices;
