--- src/server/models/paragraphContent.ts
@@ -1,5 +1,0 @@
-export class ParagraphContent {
-
-  type: string = "ParagraphContent";
-
-}
